requirements.txt
flask
requests
gspread
oauth2client
python-dotenv
sentence-transformers
